+++ themes/theme-core/src/components/Dialog.styles.ts
@@ -1,13 +1,16 @@
 import { ThemeComponent, cva } from '@marigold/system';
 
 export const Dialog: ThemeComponent<'Dialog'> = {
-  container: cva('bg-bg-surface rounded-sm px-8 pb-8 pt-4', {
-    variants: {
-      size: {
-        medium: 'w-[600px]',
+  container: cva(
+    'bg-bg-surface-overlay shadow-surface-overlay rounded-sm px-8 pb-8 pt-4',
+    {
+      variants: {
+        size: {
+          medium: 'w-[600px]',
+        },
       },
-    },
-  }),
+    }
+  ),
   closeButton: cva([
     'relative right-[-24px] top-[-8px] size-6',
     'bg-bg-brand text-text-inverted',