+++ themes/theme-core/src/components/Tag.styles.ts
@@ -2,7 +2,8 @@ import { ThemeComponent, cva } from '@marigold/system';
 
 export const Tag: ThemeComponent<'Tag'> = {
   tag: cva([
-    'border-border-base bg-bg-base flex items-center gap-1 rounded border',
+    'border-border-base bg-bg-base flex items-center gap-1 rounded border ',
+    'data-[selected]:bg-bg-selected-input data-[selected]:text-text-inverted data-[selected]:border-border-selected',
     'px-1.5 py-[2px]',
   ]),
   listItems: cva('flex flex-wrap items-center gap-1'),